convert existing js components to ts
