+++ src/components/header.tsx
@@ -1,8 +1,11 @@
+import * as React from 'react';
 import { Link } from 'gatsby';
-import PropTypes from 'prop-types';
-import React from 'react';
 
-const Header = ({ siteTitle }) => (
+interface HeaderProps {
+  siteTitle?: string;
+}
+
+const Header: React.FC<HeaderProps> = ({ siteTitle }) => (
   <header
     style={{
       background: `#2496ED`,
@@ -28,12 +31,8 @@ const Header = ({ siteTitle }) => (
   </header>
 );
 
-Header.propTypes = {
-  siteTitle: PropTypes.string,
-};
-
 Header.defaultProps = {
-  siteTitle: ``,
+  siteTitle: '',
 };
 
 export default Header;
