+++ src/components/image.tsx
@@ -1,0 +1,21 @@
+import * as React from 'react';
+import { useStaticQuery, graphql } from 'gatsby';
+import Img from 'gatsby-image';
+
+const Image: React.FC = () => {
+  const data = useStaticQuery(graphql`
+    query {
+      placeholderImage: file(relativePath: { eq: "gatsby-astronaut.png" }) {
+        childImageSharp {
+          fluid(maxWidth: 300) {
+            ...GatsbyImageSharpFluid
+          }
+        }
+      }
+    }
+  `);
+
+  return <Img fluid={data.placeholderImage.childImageSharp.fluid} />;
+};
+
+export default Image;
