+++ src/components/layout.tsx
@@ -1,18 +1,14 @@
-/**
- * Layout component that queries for data
- * with Gatsby's useStaticQuery component
- *
- * See: https://www.gatsbyjs.org/docs/use-static-query/
- */
-
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
 
 import Header from './header';
 import './layout.css';
 
-const Layout = ({ children }) => {
+interface LayoutProps {
+  children?: React.ReactNode;
+}
+
+const Layout: React.FC<LayoutProps> = ({ children }) => {
   const data = useStaticQuery(graphql`
     query SiteTitleQuery {
       site {
@@ -33,14 +29,12 @@ const Layout = ({ children }) => {
           padding: `0 1.0875rem 1.45rem`,
         }}>
         <main>{children}</main>
-        <footer>© {new Date().getFullYear()}, Codivox</footer>
+        <footer>
+          © {new Date().getFullYear()}, {data.site.siteMetadata.title}
+        </footer>
       </div>
     </>
   );
-};
-
-Layout.propTypes = {
-  children: PropTypes.node.isRequired,
 };
 
 export default Layout;
